backend/requirements.txt

# ─── Web Server ───────────────────────────
fastapi==0.115.5
uvicorn==0.30.0
websockets==12.0
python-multipart==0.0.9
pydantic==2.7.0

# ─── Google AI ────────────────────────────
google-genai
google-adk
google-cloud-aiplatform

# ─── Google Cloud Services ────────────────
google-cloud-documentai
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-auth==2.29.0
google-auth-httplib2

# ─── Browser Automation ───────────────────
playwright==1.45.0
playwright-stealth==1.0.6

# ─── PDF Processing ───────────────────────
pymupdf==1.24.0

# ─── Utilities ────────────────────────────
pillow==10.4.0
python-dotenv==1.0.0
httpx==0.27.0
aiofiles==23.2.1
sentence-transformers
pytest
pytest-asyncio
